requirements.txt
python-barcode==0.10.0
